fixes to install scripts

+++ install-ubuntu-18.sh
@@ -21,8 +21,10 @@ if [ ! -d urxvt-perls/.git ]; then
     rm -rf urxvt-perls
     git clone https://github.com/muennich/urxvt-perls urxvt-perls
     cd urxvt-perls
-    mkdir -p "$HOME/.urxvt/ext"
-    ln -s "$PWD/keyboard-select" "$HOME/.urxvt/ext/"
+    if [ ! -e "$HOME/.urxvt/ext/keyboard-select" ]; then
+        mkdir -p "$HOME/.urxvt/ext"
+        ln -s "$PWD/keyboard-select" "$HOME/.urxvt/ext/"
+    fi
 fi
 
 echo "Installing Hack Font"
